Refatora componente de Perfil para melhor legibilidade
